Updated after JMS repo split changes.
Fixed ActiveMQ 'scs_whenNoConsumerGroups_appsReceiveMessage' test.
