+++ spring-cloud-stream-binder-jms-activemq/src/test/java/org/springframework/cloud/stream/binder/jms/activemq/RepublishMessageRecovererTests.java
@@ -33,8 +33,8 @@ import org.mockito.Mockito;
 
 import org.springframework.cloud.stream.binder.jms.spi.QueueProvisioner;
 import org.springframework.cloud.stream.binder.jms.test.ActiveMQTestUtils;
-import org.springframework.integration.jms.DefaultJmsHeaderMapper;
 import org.springframework.cloud.stream.binder.jms.utils.RepublishMessageRecoverer;
+import org.springframework.integration.jms.DefaultJmsHeaderMapper;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.jms.core.MessageCreator;
 
@@ -45,9 +45,6 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.springframework.cloud.stream.binder.jms.utils.RepublishMessageRecoverer.X_EXCEPTION_MESSAGE;
-import static org.springframework.cloud.stream.binder.jms.utils.RepublishMessageRecoverer.X_EXCEPTION_STACKTRACE;
-import static org.springframework.cloud.stream.binder.jms.utils.RepublishMessageRecoverer.X_ORIGINAL_QUEUE;
 
 public class RepublishMessageRecovererTests {
 
@@ -89,7 +86,7 @@ public class RepublishMessageRecovererTests {
         target.recover(message, cause);
         message = jmsTemplate.receive(DEAD_LETTER_QUEUE);
 
-        assertThat(message.getStringProperty(X_EXCEPTION_STACKTRACE), containsString(exceptionMessage));
+        assertThat(message.getStringProperty(RepublishMessageRecoverer.X_EXCEPTION_STACKTRACE), containsString(exceptionMessage));
     }
 
     @Test
@@ -99,7 +96,7 @@ public class RepublishMessageRecovererTests {
         target.recover(message, nestedCause);
         message = jmsTemplate.receive(DEAD_LETTER_QUEUE);
 
-        assertThat(message.getStringProperty(X_EXCEPTION_MESSAGE), is(exceptionMessage));
+        assertThat(message.getStringProperty(RepublishMessageRecoverer.X_EXCEPTION_MESSAGE), is(exceptionMessage));
 
     }
 
@@ -108,7 +105,7 @@ public class RepublishMessageRecovererTests {
         target.recover(message, cause);
         message = jmsTemplate.receive(DEAD_LETTER_QUEUE);
 
-        assertThat(message.getStringProperty(X_EXCEPTION_MESSAGE), is(exceptionMessage));
+        assertThat(message.getStringProperty(RepublishMessageRecoverer.X_EXCEPTION_MESSAGE), is(exceptionMessage));
     }
 
     @Test
@@ -116,7 +113,7 @@ public class RepublishMessageRecovererTests {
         target.recover(message, cause);
         message = jmsTemplate.receive(DEAD_LETTER_QUEUE);
 
-        assertThat(message.getStringProperty(X_ORIGINAL_QUEUE), is("queue://my-fancy-queue"));
+        assertThat(message.getStringProperty(RepublishMessageRecoverer.X_ORIGINAL_QUEUE), is("queue://my-fancy-queue"));
     }
 
     @Test
