+++ spring-cloud-stream-binder-jms-common/src/main/java/org/springframework/cloud/stream/binder/jms/utils/AnonymousNamingStrategy.java
@@ -1,5 +1,8 @@
 package org.springframework.cloud.stream.binder.jms.utils;
 
+/**
+ * @author Donovan Muller
+ */
 public interface AnonymousNamingStrategy {
 
     String generateName();
