+++ spring-cloud-stream-binder-jms-common/src/main/resources/spring-configuration-metadata.json
